Consolidate audio CLI tools to pw-cat (PipeWire)
Issue #25: Update TypingSoundPlayer to use pw-cat
- Changed from pw-play to pw-cat (pw-play is just a symlink)
- Updated comments and warning messages

Issue #26: Rename AlsaAudioRecorder to PipeWireAudioRecorder
- Renamed class and file to Audio/PipeWireAudioRecorder.cs
- Changed namespace to Olbrasoft.PushToTalk.Audio
- Changed from pw-record to pw-cat -r
- Updated DI registrations in App and Service

pw-cat uses Linux kernel uinput and works on both Wayland and X11.
Fallback to paplay/arecord preserved for systems without PipeWire.

Closes #24, closes #25, closes #26
